Update to implement parent class 33500 and child class 33521A
