+++ docs/api/instruments/agilent/agilent33500.rst
@@ -1,0 +1,7 @@
+################################
+Agilent 33500 Function/Arbitrary Waveform Generator Family
+################################
+
+.. autoclass:: pymeasure.instruments.agilent.Agilent33500
+    :members:
+    :show-inheritance:
