+++ docs/api/instruments/agilent/index.rst
@@ -16,4 +16,5 @@ This section contains specific documentation on the Agilent instruments that are
    agilent34410A
    agilent4156
    agilent33220A
+   agilent33500
    agilent33521A
