+++ pymeasure/instruments/agilent/__init__.py
@@ -29,4 +29,5 @@ from .agilentE4980 import AgilentE4980
 from .agilent34410A import Agilent34410A
 from .agilent4156 import Agilent4156
 from .agilent33220A import Agilent33220A
+from .agilent33500 import Agilent33500
 from .agilent33521A import Agilent33521A
